refactor(web/hooks/chat): introduce useChatPersistence and useCopilotActions hooks
- Added `useChatPersistence` hook to manage chat message persistence and synchronization with the database, improving chat initialization and message saving logic.
- Created `useCopilotActions` hook to register various copilot actions for chat tools, enhancing the extensibility of chat functionalities.
- Updated imports in `page.tsx` to reflect the new hook structure, ensuring cleaner code organization.
